requirements.txt
langchain
langgraph
langchain-community
langchain-openrouter
python-dotenv
pydantic
duckduckgo-search
wikipedia
pathvalidate